Implemented changes requested in pr #440
